refactor: prepare wave 6 shared readiness
