+++ architecture/modules/embedding-contract.json
@@ -1,0 +1,19 @@
+{
+  "schema_version": 1,
+  "capability": "embedding-contract",
+  "completed_issues": [],
+  "go": {
+    "units": [
+      {
+        "id": "github.com/markhuangai/dense-mem/internal/embedding/contract",
+        "role": "port",
+        "visibility": "public"
+      }
+    ]
+  },
+  "browser": {
+    "units": []
+  },
+  "exceptions": [],
+  "workers": []
+}
