+++ architecture/modules/memory-application.json
@@ -2,11 +2,34 @@
   "schema_version": 1,
   "capability": "memory-application",
   "source_ownership": [
-    {"path": "cmd/internal/serverapp/recall_composition.go", "issue": 376},
-    {"path": "cmd/internal/serverapp/lifecycle_composition.go", "issue": 374},
-    {"path": "cmd/internal/serverapp/recall_feedback_composition.go", "issue": 376},
-    {"path": "internal/tools/registry/recall_bindings.go", "issue": 376},
-    {"path": "internal/tools/registry/lifecycle_bindings.go", "issue": 374}
+    {
+      "path": "cmd/internal/serverapp/recall_composition.go",
+      "issue": 376
+    },
+    {
+      "path": "cmd/internal/serverapp/lifecycle_composition.go",
+      "issue": 374
+    },
+    {
+      "path": "cmd/internal/serverapp/recall_feedback_composition.go",
+      "issue": 376
+    },
+    {
+      "path": "internal/tools/registry/recall_bindings.go",
+      "issue": 376
+    },
+    {
+      "path": "internal/tools/registry/lifecycle_bindings.go",
+      "issue": 374
+    },
+    {
+      "path": "internal/service/memoryservice/lifecycle.go",
+      "issue": 374
+    },
+    {
+      "path": "internal/service/memoryservice/lifecycle_test.go",
+      "issue": 374
+    }
   ],
   "completed_issues": [],
   "go": {
@@ -22,12 +45,6 @@
     "units": []
   },
   "exceptions": [
-    {
-      "source": "github.com/markhuangai/dense-mem/internal/service/memoryservice",
-      "target": "github.com/markhuangai/dense-mem/internal/embedding",
-      "removal_issue": 376,
-      "reason": "Recall application code still calls the embedding adapter until recall ownership is isolated."
-    },
     {
       "source": "github.com/markhuangai/dense-mem/internal/service/memoryservice",
       "target": "github.com/markhuangai/dense-mem/internal/repository",
