+++ internal/repository/source_revision_repository.go
@@ -48,33 +48,3 @@ func sourceKindForEvidence(sourceType string) string {
 		return "conversation"
 	}
 }
-
-// withSystemModeInTx is a shared RLS transition used by legacy adapters that
-// are outside the knowledge write-owner migration.
-func withSystemModeInTx(ctx context.Context, tx *gorm.DB, teamID, profileID string, fn func(systemTx *gorm.DB) error) error {
-	if err := tx.WithContext(ctx).Exec("SELECT set_config('app.current_team_id', '', true)").Error; err != nil {
-		return err
-	}
-	if err := tx.WithContext(ctx).Exec("SELECT set_config('app.current_profile_id', '', true)").Error; err != nil {
-		return err
-	}
-	if err := tx.WithContext(ctx).Exec("SELECT set_config('app.tx_mode', 'system', true)").Error; err != nil {
-		return err
-	}
-	fnErr := fn(tx)
-	resetErr := resetProfileModeInTx(ctx, tx, teamID, profileID)
-	if fnErr != nil {
-		return fnErr
-	}
-	return resetErr
-}
-
-func resetProfileModeInTx(ctx context.Context, tx *gorm.DB, teamID, profileID string) error {
-	if err := tx.WithContext(ctx).Exec("SELECT set_config('app.current_team_id', ?, true)", teamID).Error; err != nil {
-		return err
-	}
-	if err := tx.WithContext(ctx).Exec("SELECT set_config('app.current_profile_id', ?, true)", profileID).Error; err != nil {
-		return err
-	}
-	return tx.WithContext(ctx).Exec("SELECT set_config('app.tx_mode', 'profile', true)").Error
-}
