+++ scripts/e2e-db-cases/lifecycle.json
@@ -1,5 +1,5 @@
 {
   "version": 1,
-  "capability": "server",
+  "capability": "lifecycle",
   "cases": []
 }
