+++ scripts/e2e-db-cases/memorypack.json
@@ -1,5 +1,5 @@
 {
   "version": 1,
-  "capability": "server",
+  "capability": "memorypack",
   "cases": []
 }
