+++ scripts/e2e-db-cases/service.json
@@ -1,63 +1,5 @@
 {
   "version": 1,
   "capability": "service",
-  "cases": [
-    {
-      "id": "service/TestComposeSynchronousEvidenceOnlyAssessorBatch",
-      "package": "./internal/service/memoryservice",
-      "run": "^TestComposeSynchronousEvidenceOnlyAssessorBatch$",
-      "phase": "scenario",
-      "scenario": "synchronous_write_primitives",
-      "source": "internal/service/memoryservice/synchronous_assessment_compose_e2e.e2e"
-    },
-    {
-      "id": "service/TestTeamServiceCreate",
-      "package": "./internal/service",
-      "run": "^TestTeamServiceCreate$",
-      "phase": "precheck",
-      "source": "internal/service/team_service.e2e"
-    },
-    {
-      "id": "service/TestTeamServiceCreateDuplicateName",
-      "package": "./internal/service",
-      "run": "^TestTeamServiceCreateDuplicateName$",
-      "phase": "precheck",
-      "source": "internal/service/team_service.e2e"
-    },
-    {
-      "id": "service/TestTeamServiceDelete",
-      "package": "./internal/service",
-      "run": "^TestTeamServiceDelete$",
-      "phase": "precheck",
-      "source": "internal/service/team_service.e2e"
-    },
-    {
-      "id": "service/TestTeamServiceDeleteRevokesActiveKeys",
-      "package": "./internal/service",
-      "run": "^TestTeamServiceDeleteRevokesActiveKeys$",
-      "phase": "precheck",
-      "source": "internal/service/team_service.e2e"
-    },
-    {
-      "id": "service/TestTeamServiceGet",
-      "package": "./internal/service",
-      "run": "^TestTeamServiceGet$",
-      "phase": "precheck",
-      "source": "internal/service/team_service.e2e"
-    },
-    {
-      "id": "service/TestTeamServiceList",
-      "package": "./internal/service",
-      "run": "^TestTeamServiceList$",
-      "phase": "precheck",
-      "source": "internal/service/team_service.e2e"
-    },
-    {
-      "id": "service/TestTeamServiceUpdate",
-      "package": "./internal/service",
-      "run": "^TestTeamServiceUpdate$",
-      "phase": "precheck",
-      "source": "internal/service/team_service.e2e"
-    }
-  ]
+  "cases": []
 }
